Final submission cleanup: remove redundant docs/reports, streamline Qwen2.5-1.5B F16/Q8/Q4 local setup, and validate all tests

+++ tests/test_cliff.py
@@ -35,7 +35,7 @@ def test_both_conditions_required_practical_only_is_not_a_cliff(criterion):
     """A 20pp difference at n=10 meets the threshold but the CI straddles zero."""
     result = analyse_metric(
         "ps3", "m", "higher_is_better", 0.05,
-        arms(bf16=(9, 10), fp8=(9, 10), q8=(8, 10), q4=(7, 10)), criterion,
+        arms(f16=(9, 10), fp8=(9, 10), q8=(8, 10), q4=(7, 10)), criterion,
     )
     q4 = next(p for p in result.points if p.precision == "q4")
     assert q4.meets_practical_threshold is True
@@ -49,7 +49,7 @@ def test_both_conditions_required_statistical_only_is_not_a_cliff(criterion):
     """A tiny difference that is statistically clean at huge n is not operationally a cliff."""
     result = analyse_metric(
         "ps3", "m", "higher_is_better", 0.05,
-        arms(bf16=(9900, 10000), fp8=(9900, 10000), q8=(9900, 10000), q4=(9750, 10000)),
+        arms(f16=(9900, 10000), fp8=(9900, 10000), q8=(9900, 10000), q4=(9750, 10000)),
         criterion,
     )
     q4 = next(p for p in result.points if p.precision == "q4")
@@ -61,7 +61,7 @@ def test_both_conditions_required_statistical_only_is_not_a_cliff(criterion):
 def test_both_conditions_met_is_a_cliff(criterion):
     result = analyse_metric(
         "ps3", "m", "higher_is_better", 0.05,
-        arms(bf16=(950, 1000), fp8=(948, 1000), q8=(945, 1000), q4=(600, 1000)), criterion,
+        arms(f16=(950, 1000), fp8=(948, 1000), q8=(945, 1000), q4=(600, 1000)), criterion,
     )
     q4 = next(p for p in result.points if p.precision == "q4")
     assert q4.past_cliff is True
@@ -77,7 +77,7 @@ def test_lower_is_better_orientation(criterion):
     """For violation_rate, a HIGHER value is worse and must yield positive degradation."""
     result = analyse_metric(
         "ps1", "m", "lower_is_better", 0.02,
-        arms(bf16=(20, 1000), fp8=(22, 1000), q8=(25, 1000), q4=(200, 1000)), criterion,
+        arms(f16=(20, 1000), fp8=(22, 1000), q8=(25, 1000), q4=(200, 1000)), criterion,
     )
     q4 = next(p for p in result.points if p.precision == "q4")
     assert q4.degradation > 0
@@ -87,7 +87,7 @@ def test_lower_is_better_orientation(criterion):
 def test_improvement_yields_negative_degradation(criterion):
     result = analyse_metric(
         "ps3", "m", "higher_is_better", 0.05,
-        arms(bf16=(800, 1000), q4=(900, 1000)), criterion,
+        arms(f16=(800, 1000), q4=(900, 1000)), criterion,
     )
     q4 = next(p for p in result.points if p.precision == "q4")
     assert q4.degradation < 0
@@ -102,7 +102,7 @@ def test_gradual_slope_is_not_called_a_cliff(criterion):
     """Even steps of ~10pp each: the criterion fires, but nothing dominates."""
     result = analyse_metric(
         "ps3", "m", "higher_is_better", 0.05,
-        arms(bf16=(950, 1000), fp8=(850, 1000), q8=(750, 1000), q4=(650, 1000)), criterion,
+        arms(f16=(950, 1000), fp8=(850, 1000), q8=(750, 1000), q4=(650, 1000)), criterion,
     )
     assert result.cliff_precision == "fp8"
     assert result.pattern == "gradual"
@@ -112,7 +112,7 @@ def test_gradual_slope_is_not_called_a_cliff(criterion):
 def test_single_dominant_step_is_a_cliff(criterion):
     result = analyse_metric(
         "ps3", "m", "higher_is_better", 0.05,
-        arms(bf16=(950, 1000), fp8=(948, 1000), q8=(946, 1000), q4=(500, 1000)), criterion,
+        arms(f16=(950, 1000), fp8=(948, 1000), q8=(946, 1000), q4=(500, 1000)), criterion,
     )
     assert result.pattern == "cliff"
     assert result.dominant_step["from"] == "q8" and result.dominant_step["to"] == "q4"
@@ -125,7 +125,7 @@ def test_single_dominant_step_is_a_cliff(criterion):
 def test_no_degradation_reports_minimum_detectable_difference(criterion):
     result = analyse_metric(
         "ps3", "m", "higher_is_better", 0.05,
-        arms(bf16=(90, 100), fp8=(90, 100), q8=(90, 100), q4=(90, 100)), criterion,
+        arms(f16=(90, 100), fp8=(90, 100), q8=(90, 100), q4=(90, 100)), criterion,
     )
     assert result.pattern == "none"
     assert result.cliff_precision is None
@@ -146,7 +146,7 @@ def test_missing_reference_arm_is_refused_not_substituted(criterion):
 def test_small_sample_is_noted(criterion):
     result = analyse_metric(
         "ps1", "m", "lower_is_better", 0.02,
-        arms(bf16=(1, 10), q4=(5, 10)), criterion,
+        arms(f16=(1, 10), q4=(5, 10)), criterion,
     )
     assert any("small-sample" in n for n in result.notes)
 
@@ -159,11 +159,11 @@ def test_worst_metric_governs_across_suites(criterion):
     """Safety fine at Q4 but tool-calling broken -> Q4 must not be recommended."""
     per_suite = {
         "ps1": {p: {"violation_rate": blob(k, n), "benign_refusal_rate": blob(1, 100)}
-                for p, (k, n) in [("bf16", (20, 1000)), ("fp8", (20, 1000)),
+                for p, (k, n) in [("f16", (20, 1000)), ("fp8", (20, 1000)),
                                   ("q8", (21, 1000)), ("q4", (22, 1000))]},
         "ps3": {p: {"task_success_rate": blob(k, n),
                     "structured_output_validity": blob(990, 1000)}
-                for p, (k, n) in [("bf16", (950, 1000)), ("fp8", (948, 1000)),
+                for p, (k, n) in [("f16", (950, 1000)), ("fp8", (948, 1000)),
                                   ("q8", (945, 1000)), ("q4", (500, 1000))]},
     }
     out = detect_cliffs(per_suite, criterion)
@@ -177,11 +177,11 @@ def test_safety_failure_alone_rejects_a_precision(criterion):
     """The mirror case: tool-calling fine at Q4, safety collapsed."""
     per_suite = {
         "ps1": {p: {"violation_rate": blob(k, n), "benign_refusal_rate": blob(1, 100)}
-                for p, (k, n) in [("bf16", (20, 1000)), ("fp8", (21, 1000)),
+                for p, (k, n) in [("f16", (20, 1000)), ("fp8", (21, 1000)),
                                   ("q8", (22, 1000)), ("q4", (300, 1000))]},
         "ps3": {p: {"task_success_rate": blob(950, 1000),
                     "structured_output_validity": blob(990, 1000)}
-                for p in ["bf16", "fp8", "q8", "q4"]},
+                for p in ["f16", "fp8", "q8", "q4"]},
     }
     out = detect_cliffs(per_suite, criterion)
     assert out["minimum_viable_precision"]["precision"] == "q8"
@@ -190,10 +190,10 @@ def test_safety_failure_alone_rejects_a_precision(criterion):
 def test_no_cliff_anywhere_permits_the_lowest_tested_precision(criterion):
     per_suite = {
         "ps1": {p: {"violation_rate": blob(20, 1000), "benign_refusal_rate": blob(1, 100)}
-                for p in ["bf16", "fp8", "q8", "q4"]},
+                for p in ["f16", "fp8", "q8", "q4"]},
         "ps3": {p: {"task_success_rate": blob(950, 1000),
                     "structured_output_validity": blob(990, 1000)}
-                for p in ["bf16", "fp8", "q8", "q4"]},
+                for p in ["f16", "fp8", "q8", "q4"]},
     }
     out = detect_cliffs(per_suite, criterion)
     mvp = out["minimum_viable_precision"]
@@ -202,7 +202,7 @@ def test_no_cliff_anywhere_permits_the_lowest_tested_precision(criterion):
 
 
 def test_mvp_always_carries_the_scope_caveat(criterion):
-    out = detect_cliffs({"ps1": {"bf16": {"violation_rate": blob(1, 100),
+    out = detect_cliffs({"ps1": {"f16": {"violation_rate": blob(1, 100),
                                           "benign_refusal_rate": blob(1, 100)}}}, criterion)
     assert "universally safe production precision" in out["minimum_viable_precision"]["caveat"]
